+++ GameState.py
@@ -1,0 +1,56 @@
+from dataclasses import dataclass
+from enum import Enum, auto
+from typing import List, Tuple
+from collections import Counter
+from Card import Card
+from Players import Player
+
+
+@dataclass
+class GamePhase(Enum):
+    ANTE = auto()
+    DEALING = auto()
+    BETTING = auto()
+    FOLDING = auto()
+    DRAW = auto()
+    SHOWDOWN = auto()
+
+@dataclass
+class GameState:
+    deck: List[Card]      # Defines a property 'deck' of type List[Card]. This will store the deck of cards in the game.
+    players: List[Player] # Defines a property 'players' of type List[List[Card]]. This will store the hands of all players.
+    gamePhase: GamePhase
+    pot: int              # Integer value representing the total chips in the pot
+
+    def RankHand(hand: List[Card]) -> Tuple[int, List[int]]:
+        ranks = [card.rank.value for card in hand]
+        if set(ranks) == {0, 10, 11, 12, 13}:                 # Convert placeholder back to 0 for Page
+            ranks.remove(0)
+            ranks.append(-1)                                  # Add a placeholder for Page (to be converted to 0 later)
+        ranks.sort(reverse=True)                              # Sort ranks in descending order
+        ranks = [rank if rank != -1 else 0 for rank in ranks] # Convert placeholder back to 0 for Page
+        suits = [card.suit for card in hand]
+        rankCounts = Counter(ranks)
+        isFlush = len(set(suits)) == 1
+        isStraight = len(set(ranks)) == 5 and max(ranks) - min(ranks) == 4
+        if isFlush and isStraight:
+            if max(ranks) == 14:
+                return (10, ranks) # Royal Flush
+            else:
+                return (9, ranks)  # Straight Flush
+        elif max(rankCounts.values()) == 4:
+            return (8, ranks)      # Four of a Kind
+        elif max(rankCounts.values()) == 3 and len(rankCounts) == 2:
+            return (7, ranks)      # Full House
+        elif isFlush:
+            return (6, ranks)      # Flush
+        elif isStraight:
+            return (5, ranks)      # Straight
+        elif max(rankCounts.values()) == 3:
+            return (4, ranks)      # Three of a Kind
+        elif list(rankCounts.values()).count(2) == 2:
+            return (3, ranks)      # Two Pair
+        elif list(rankCounts.values()).count(2) == 1:
+            return (2, ranks)      # One Pair
+        else:
+            return (1, ranks)      # High Card