+++ changelog.md
@@ -7,6 +7,8 @@ The formatting is based on *[Keep a Changelog](https://keepachangelog.com/en/1.0
 and this project adheres to *[Semantic Versioning](https://semver.org/spec/v2.0.0.html)*.
 
 ## [Unreleased]
+- Added AIDealer class / Nathan 2023-10-28
+- Change file structure to split AIDealer.py and GameState.py with related classes / Nathan 2023-10-28
 - Added ShuffleAllToDeck(), Player class, changelog.md, changelogtemplate.md / Nathan 2023-10-27
 - Change file structure to split Card.py and Players.py with related classes / Nathan 2023-10-27
 - Added game phases / Nathan 2023-10-26
@@ -15,4 +17,4 @@ and this project adheres to *[Semantic Versioning](https://semver.org/spec/v2.0.
 ## [0.0.1] - 2023-10-26
 ### Added
 - Protype in python / Nathan 2023-10-25
-- readme.md
+- readme.md / Nathan 2023-10-26